update computation and plotting
